+++ test/framework/src/main/java/org/opensearch/cluster/coordination/AbstractCoordinatorTestCase.java
@@ -1017,6 +1017,11 @@ public class AbstractCoordinatorTestCase extends OpenSearchTestCase {
             }
 
             @Override
+            public PersistedStateStats getStats() {
+                return null;
+            }
+
+            @Override
             public void close() {
                 assertTrue(openPersistedStates.remove(this));
                 try {